requirements.txt
pytest==7.4.0
selenium==4.10.0
ipython==8.14.0
python-dotenv==1.0.0
allure-pytest==2.13.2